// src/models/car-models/types/CarModelIsSuccessorOfRelationship.ts
import type {BaseNode} from "../../../controllers/nodes/types/BaseNode"

export type CarModelIsSuccessorOfRelationship = {
    car_model_id: number
    partner_id: number
    relationship_id: number
    relationship_name: string
    relationship_partner?: BaseNode
    created_at: string
    updated_at: string
}